Delete SpanSummaryPanel
